make scan_ipc work on s3
