+++ py-polars/polars/__init__.py
@@ -58,6 +58,7 @@ from polars.internals.frame import (  # flake8: noqa # TODO: remove need for wra
     wrap_df,
 )
 from polars.internals.functions import arg_where, concat, date_range, get_dummies
+from polars.internals.io import read_ipc_schema
 from polars.internals.lazy_frame import LazyFrame
 from polars.internals.lazy_functions import _date as date
 from polars.internals.lazy_functions import _datetime as datetime
@@ -113,7 +114,6 @@ from polars.io import (
     read_csv,
     read_excel,
     read_ipc,
-    read_ipc_schema,
     read_json,
     read_parquet,
     read_sql,
