+++ py-polars/polars/internals/anonymous_scan.py
@@ -1,6 +1,6 @@
 import pickle
 from functools import partial
-from typing import List, Optional
+from typing import Dict, List, Optional
 
 import polars as pl
 from polars import internals as pli
@@ -64,3 +64,31 @@ def _scan_ds(ds: "pa.dataset.dataset") -> "pli.LazyFrame":
     func = partial(_scan_ds_impl, ds)
     func_serialized = pickle.dumps(func)
     return pli.LazyFrame._scan_python_function(ds.schema, func_serialized)
+
+
+def _scan_ipc_impl(uri: "str", with_columns: Optional[List[str]]) -> "pli.DataFrame":
+    """
+    Takes the projected columns and materializes an arrow table.
+
+    Parameters
+    ----------
+    uri
+    with_columns
+    """
+    import polars as pl
+
+    return pl.read_ipc(uri, with_columns)
+
+
+def _scan_ipc_fsspec(
+    file: str,
+    storage_options: Optional[Dict] = None,
+) -> "pli.LazyFrame":
+    func = partial(_scan_ipc_impl, file)
+    func_serialized = pickle.dumps(func)
+
+    storage_options = storage_options or {}
+    with pli._prepare_file_arg(file, **storage_options) as data:
+        schema = pli.read_ipc_schema(data)
+
+    return pli.LazyFrame._scan_python_function(schema, func_serialized)
