+++ py-polars/polars/io.py
@@ -1,13 +1,11 @@
-from contextlib import contextmanager
+import os.path
 from io import BytesIO, IOBase, StringIO
 from pathlib import Path
 from typing import (
     Any,
     BinaryIO,
     Callable,
-    ContextManager,
     Dict,
-    Iterator,
     List,
     Mapping,
     Optional,
@@ -16,9 +14,7 @@ from typing import (
     Type,
     Union,
     cast,
-    overload,
 )
-from urllib.request import urlopen
 
 from polars.utils import format_path, handle_projection_columns
 
@@ -34,12 +30,8 @@ except ImportError:  # pragma: no cover
 
 from polars.convert import from_arrow
 from polars.datatypes import DataType
-from polars.internals import DataFrame, LazyFrame, _scan_ds
-
-try:
-    from polars.polars import ipc_schema as _ipc_schema
-except ImportError:  # pragma: no cover
-    pass
+from polars.internals import DataFrame, LazyFrame, _scan_ds, _scan_ipc_fsspec
+from polars.internals.io import _prepare_file_arg
 
 try:
     import connectorx as cx
@@ -47,14 +39,6 @@ try:
     _WITH_CX = True
 except ImportError:
     _WITH_CX = False
-
-try:
-    import fsspec
-    from fsspec.utils import infer_storage_options
-
-    _WITH_FSSPEC = True
-except ImportError:
-    _WITH_FSSPEC = False
 
 
 def _check_arg_is_1byte(
@@ -71,78 +55,6 @@ def _check_arg_is_1byte(
             raise ValueError(
                 f'{arg_name}="{arg}" should be a single byte character, but is {arg_byte_length} bytes long.'
             )
-
-
-def _process_http_file(path: str) -> BytesIO:
-    with urlopen(path) as f:
-        return BytesIO(f.read())
-
-
-@overload
-def _prepare_file_arg(
-    file: Union[str, List[str], Path, BinaryIO, bytes], **kwargs: Any
-) -> ContextManager[Union[str, BinaryIO]]:
-    ...
-
-
-@overload
-def _prepare_file_arg(
-    file: Union[str, TextIO, Path, BinaryIO, bytes], **kwargs: Any
-) -> ContextManager[Union[str, BinaryIO]]:
-    ...
-
-
-@overload
-def _prepare_file_arg(
-    file: Union[str, List[str], TextIO, Path, BinaryIO, bytes], **kwargs: Any
-) -> ContextManager[Union[str, List[str], BinaryIO, List[BinaryIO]]]:
-    ...
-
-
-def _prepare_file_arg(
-    file: Union[str, List[str], TextIO, Path, BinaryIO, bytes], **kwargs: Any
-) -> ContextManager[Union[str, BinaryIO, List[str], List[BinaryIO]]]:
-    """
-    Utility for read_[csv, parquet]. (not to be used by scan_[csv, parquet]).
-    Returned value is always usable as a context.
-
-    A `StringIO`, `BytesIO` file is returned as a `BytesIO`.
-    A local path is returned as a string.
-    An http URL is read into a buffer and returned as a `BytesIO`.
-
-    When fsspec is installed, remote file(s) is (are) opened with
-    `fsspec.open(file, **kwargs)` or `fsspec.open_files(file, **kwargs)`.
-    """
-
-    # Small helper to use a variable as context
-    @contextmanager
-    def managed_file(file: Any) -> Iterator[Any]:
-        try:
-            yield file
-        finally:
-            pass
-
-    if isinstance(file, StringIO):
-        return BytesIO(file.read().encode("utf8"))
-    if isinstance(file, BytesIO):
-        return managed_file(file)
-    if isinstance(file, Path):
-        return managed_file(format_path(file))
-    if isinstance(file, str):
-        if _WITH_FSSPEC:
-            if infer_storage_options(file)["protocol"] == "file":
-                return managed_file(format_path(file))
-            return fsspec.open(file, **kwargs)
-        if file.startswith("http"):
-            return _process_http_file(file)
-    if isinstance(file, list) and bool(file) and all(isinstance(f, str) for f in file):
-        if _WITH_FSSPEC:
-            if all(infer_storage_options(f)["protocol"] == "file" for f in file):
-                return managed_file([format_path(f) for f in file])
-            return fsspec.open_files(file, **kwargs)
-    if isinstance(file, str):
-        file = format_path(file)
-    return managed_file(file)
 
 
 def update_columns(df: DataFrame, new_columns: List[str]) -> DataFrame:
@@ -623,6 +535,7 @@ def scan_ipc(
     rechunk: bool = True,
     row_count_name: Optional[str] = None,
     row_count_offset: int = 0,
+    storage_options: Optional[Dict] = None,
     **kwargs: Any,
 ) -> LazyFrame:
     """
@@ -652,6 +565,10 @@ def scan_ipc(
 
     if isinstance(file, (str, Path)):
         file = format_path(file)
+
+    # try fsspec scanner
+    if not os.path.exists((file)):
+        return _scan_ipc_fsspec(file, storage_options)
 
     return LazyFrame.scan_ipc(
         file=file,
@@ -712,27 +629,6 @@ def scan_parquet(
         row_count_name=row_count_name,
         row_count_offset=row_count_offset,
     )
-
-
-def read_ipc_schema(
-    file: Union[str, BinaryIO, Path, bytes]
-) -> Dict[str, Type[DataType]]:
-    """
-    Get a schema of the IPC file without reading data.
-
-    Parameters
-    ----------
-    file
-        Path to a file or a file-like object.
-
-    Returns
-    -------
-    Dictionary mapping column names to datatypes
-    """
-    if isinstance(file, (str, Path)):
-        file = format_path(file)
-
-    return _ipc_schema(file)
 
 
 def read_avro(
